added new landing page
