+++ client/src/components/Game/styles.scss
@@ -5,7 +5,7 @@
 
 @media (min-width: 992px) {
   .grid {
-    grid-template-columns: auto minmax(40%, auto);
+    grid-template-columns: auto minmax(30%, auto);
     grid-column-gap: 40px;
   }
