+++ client/src/components/Lobby/styles.scss
@@ -1,22 +1,32 @@
-.set-container {
+.grid-container {
   display: grid;
-  grid-gap: 10px;
-  grid-auto-flow: dense;
-  grid-template-columns: repeat(auto-fit, minmax(70px, 1fr));
 
-  .set {
-    white-space: unset;
+  .text-smaller {
+    font-size: 0.8em;
+  }
+
+  @media (min-width: 992px) {
+    flex-direction: column;
+    grid-template-columns: 1fr 1fr;
+
+    .actionable-container {
+      height: 100vh;
+      overflow-y: scroll;
+    }
+
+    .image-container {
+      height: 100vh;
+    }
+  }
+
+  .image-container {
+    overflow: hidden;
     display: flex;
     justify-content: center;
     align-items: center;
-    cursor: pointer;
-    border-width: 1px;
-    border-style: solid;
-    padding: 0.7em 1.5em;
-    box-shadow: pink;
+    img {
+      max-width: 100%;
+      height: auto;
+    }
   }
-}
-
-.language {
-  font-size: 2em;
 }
